+++ src/app/commands/commands.component.ts
@@ -9,14 +9,16 @@ import { faPowerOff, faTrash, faLockOpen, faArrowRotateLeft, faDownload } from '
 
 // other
 import { NGXLogger } from "ngx-logger";
+import { environment } from '../../environments/environment';
 
 @Component({
-  selector: 'app-commands',
-  templateUrl: './commands.component.html',
-  styleUrls: ['./commands.component.scss']
+    selector: 'app-commands',
+    templateUrl: './commands.component.html',
+    styleUrls: ['./commands.component.scss'],
+    standalone: false
 })
 export class CommandsComponent implements OnInit {
-  debug: boolean = true;
+  debug: boolean;
   private logID: string = "CommandsComponent.";
   // button command loading
   loadingResPalCode: boolean = false;
@@ -51,6 +53,11 @@ export class CommandsComponent implements OnInit {
     private notificationService: NotificationService,
     private logger: NGXLogger,
   ) {
+    if (environment.production) {
+      this.debug = false;
+    } else {
+      this.debug = true;
+    }
   }
 
   ngOnInit() {