+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Netplan GUI Changelog
 
+## 1.3.0 (7/5/2025) - [Issue #12](https://github.com/xinthose/Netplan-GUI/issues/12))
+
+- update packages (Angular 20)
+
 ## 1.2.0 (11/24/2023) - [Issue #11](https://github.com/xinthose/Netplan-GUI/issues/11))
 
 - update packages (Angular 17)
@@ -47,4 +51,4 @@
 
 ## 1.0.0 (9/26/2021)
 
-- project converted from kendo mobile (jQuery) to Angular
+- project converted from kendo mobile (jQuery) to Angular 12